refactor: extract shared force-push E2E fixtures and document unused action
1. Extract duplicated mockPR, mockFiles, config, setupForcePushScenario,
   and beforeEach setup from both collapsed-iterations spec files into
   e2e/fixtures/force-push-helpers.ts. Both specs now import shared
   helpers, eliminating ~100 lines of duplication per file.

2. Add JSDoc comment to clearCollapsedGroup action explaining it exists
   for future "dismiss without expanding" UX (no current callers).
