+++ e2e/mock/stateless-mode/collapsed-iterations-expanded.spec.ts
@@ -1,137 +1,13 @@
 import { test, expect } from "@playwright/test";
 import {
-  setupAuthState,
-  setupFullPRMocks,
-  setupStatelessIterationMocks,
-  type MockPR,
-  type MockFile,
-} from "../../fixtures/github-mocks";
-import { setupLegacyDefaults } from "../../fixtures/legacy-defaults";
+  forcePushConfig,
+  setupForcePushDefaults,
+  setupForcePushScenario,
+} from "../../fixtures/force-push-helpers";
 
 test.describe("Collapsed Iterations Expanded View", () => {
-  const mockPR: MockPR = {
-    id: 1,
-    number: 123,
-    title: "Test PR for Expanded Collapsed Groups",
-    body: "Testing expanded collapsed iteration groups",
-    state: "open",
-    merged: false,
-    draft: false,
-    user: {
-      id: 1,
-      login: "testuser",
-      avatar_url: "https://avatars.githubusercontent.com/u/1",
-    },
-    head: { ref: "feature/test", sha: "abc123" },
-    base: { ref: "main", sha: "def456" },
-    html_url: "https://github.com/test/repo/pull/123",
-    created_at: "2024-01-01T10:00:00Z",
-    updated_at: "2024-01-05T15:00:00Z",
-  };
-
-  const mockFiles: MockFile[] = [
-    {
-      filename: "src/test.ts",
-      status: "modified",
-      additions: 10,
-      deletions: 5,
-      changes: 15,
-      patch: "@@ -1,5 +1,10 @@\n+// New code\n const x = 1;",
-    },
-  ];
-
-  const config = {
-    owner: "test",
-    repo: "repo",
-    prNumber: 123,
-    pageUrl: "/test/repo/123",
-  };
-
-  /**
-   * Helper: set up a force-push scenario with mock compare API.
-   */
-  async function setupForcePushScenario(
-    page: import("@playwright/test").Page,
-    options: {
-      eventId: number;
-      beforeSha: string;
-      afterSha: string;
-      liveCommits: {
-        sha: string;
-        message: string;
-        author: string;
-        date: string;
-      }[];
-      compareResponse:
-        | {
-            status: 200;
-            commits: {
-              sha: string;
-              message: string;
-              authorName: string;
-              authorLogin: string;
-              date: string;
-            }[];
-          }
-        | { status: 404 };
-    }
-  ): Promise<void> {
-    await setupStatelessIterationMocks(
-      page,
-      config.owner,
-      config.repo,
-      config.prNumber,
-      {
-        commits: options.liveCommits,
-        timeline: [
-          {
-            id: options.eventId,
-            event: "head_ref_force_pushed",
-            created_at: "2024-01-02T12:00:00Z",
-            before_commit: { sha: options.beforeSha },
-            after_commit: { sha: options.afterSha },
-          },
-        ],
-      }
-    );
-
-    await page.route(
-      `https://api.github.com/repos/${config.owner}/${config.repo}/compare/${options.afterSha}...${options.beforeSha}`,
-      async (route) => {
-        if (options.compareResponse.status === 404) {
-          await route.fulfill({
-            status: 404,
-            contentType: "application/json",
-            body: JSON.stringify({ message: "Not Found" }),
-          });
-          return;
-        }
-
-        await route.fulfill({
-          status: 200,
-          contentType: "application/json",
-          body: JSON.stringify({
-            commits: options.compareResponse.commits.map((c) => ({
-              sha: c.sha,
-              commit: {
-                message: c.message,
-                author: { name: c.authorName, date: c.date },
-              },
-              author: { login: c.authorLogin },
-            })),
-          }),
-        });
-      }
-    );
-  }
-
   test.beforeEach(async ({ page }) => {
-    await setupLegacyDefaults(page);
-    await setupAuthState(page);
-    await setupFullPRMocks(page, config.owner, config.repo, config.prNumber, {
-      pr: mockPR,
-      files: mockFiles,
-    });
+    await setupForcePushDefaults(page);
   });
 
   test("Include button expands collapsed group into individual discarded iteration tabs", async ({
@@ -171,7 +47,7 @@ test.describe("Collapsed Iterations Expanded View", () => {
       },
     });
 
-    await page.goto(config.pageUrl);
+    await page.goto(forcePushConfig.pageUrl);
 
     // Verify initial state: collapsed group tab visible
     const collapsedTab = page.getByTestId("collapsed-group-5001");
@@ -246,7 +122,7 @@ test.describe("Collapsed Iterations Expanded View", () => {
       },
     });
 
-    await page.goto(config.pageUrl);
+    await page.goto(forcePushConfig.pageUrl);
 
     // Expand the collapsed group
     const collapsedTab = page.getByTestId("collapsed-group-6001");
@@ -327,7 +203,7 @@ test.describe("Collapsed Iterations Expanded View", () => {
       },
     });
 
-    await page.goto(config.pageUrl);
+    await page.goto(forcePushConfig.pageUrl);
 
     // Wait for iterations to load
     const selector = page.getByTestId("iteration-selector");
@@ -335,32 +211,11 @@ test.describe("Collapsed Iterations Expanded View", () => {
     const collapsedTab = page.getByTestId("collapsed-group-7001");
     await expect(collapsedTab).toBeVisible();
 
-    // Inject 'unavailable' status on the second discarded commit via store mutation
-    await page.evaluate(() => {
-      // Access the Zustand store's internal state
-      const storeState = (window as unknown as { [key: string]: unknown }).__ITERATION_STORE__;
-      if (storeState) {
-        // Direct store access — fallback below if not exposed
-        return;
-      }
-
-      // Access via localStorage-based store inspection isn't possible for non-persisted state.
-      // Instead, we'll modify collapsedGroups directly through the Zustand devtools or
-      // the store's setState. The store is available on useIterationStore.
-    });
-
-    // Since direct store mutation from E2E is fragile, we test the behavior
-    // at the rendering level: after expanding, verify that the IterationTab
-    // component correctly applies .unavailable class when isUnavailable is true.
-    //
-    // The real test: expand the group, then use page.evaluate to flip the commit
-    // status in the collapsedGroups array and trigger a re-render.
-
-    // Step 1: Expand the group
+    // Expand the group
     await collapsedTab.click();
     await page.getByTestId("collapsed-history-include-btn").click();
 
-    // Step 2: Both discarded tabs should be visible (both are 'available' from API)
+    // Both discarded tabs should be visible (both are 'available' from API)
     const discardedTab1 = page.getByTestId("iteration-tab-1");
     const discardedTab2 = page.getByTestId("iteration-tab-2");
     await expect(discardedTab1).toBeVisible();
@@ -368,30 +223,8 @@ test.describe("Collapsed Iterations Expanded View", () => {
     await expect(discardedTab1).toHaveClass(/discarded/);
     await expect(discardedTab2).toHaveClass(/discarded/);
 
-    // Step 3: Mutate store to mark second commit as unavailable and trigger re-render
-    await page.evaluate(() => {
-      // Zustand stores expose getState/setState on the hook function
-      // We access it through the module system via window.__ZUSTAND_STORES__ if available,
-      // or through the React fiber tree. For E2E, the most reliable approach is
-      // to use the store's persist key to check if the state can be modified.
-
-      // Find the iteration store in Zustand's internal registry
-      const stores = (window as unknown as { [key: string]: unknown }).__ZUSTAND_DEVTOOLS_STORES__ as
-        | Map<string, { getState: () => { [key: string]: unknown }; setState: (s: { [key: string]: unknown }) => void }>
-        | undefined;
-
-      if (!stores) {
-        // Zustand devtools not available — try direct DOM approach
-        // We'll use a different strategy: dispatch a custom event that the app can listen to
-        return;
-      }
-    });
-
-    // Since we can't reliably mutate Zustand store from E2E page context,
-    // verify the CSS class exists and has the right properties instead.
-    // This ensures the implementer added the CSS rule even if we can't trigger it via API mocks.
+    // Verify .iteration-tab.unavailable CSS rule exists with correct properties
     const unavailableStyles = await page.evaluate(() => {
-      // Check that .iteration-tab.unavailable CSS rule exists in the stylesheet
       for (const sheet of document.styleSheets) {
         try {
           for (const rule of sheet.cssRules) {
@@ -410,7 +243,6 @@ test.describe("Collapsed Iterations Expanded View", () => {
       return null;
     });
 
-    // Verify .iteration-tab.unavailable CSS rule exists with correct properties
     expect(unavailableStyles).not.toBeNull();
     expect(unavailableStyles?.opacity).toBe("0.4");
     expect(unavailableStyles?.cursor).toBe("not-allowed");
@@ -446,7 +278,7 @@ test.describe("Collapsed Iterations Expanded View", () => {
       },
     });
 
-    await page.goto(config.pageUrl);
+    await page.goto(forcePushConfig.pageUrl);
 
     // Expand the group
     const collapsedTab = page.getByTestId("collapsed-group-8001");
@@ -503,7 +335,7 @@ test.describe("Collapsed Iterations Expanded View", () => {
       },
     });
 
-    await page.goto(config.pageUrl);
+    await page.goto(forcePushConfig.pageUrl);
 
     // Expand
     const collapsedTab = page.getByTestId("collapsed-group-9001");
