+++ src/features/iterations/types.ts
@@ -213,6 +213,7 @@ export interface IterationState {
   selectRange: (fromSnapshot: number, toSnapshot: number) => void;
   selectPreset: (preset: IterationPreset) => void;
   selectCollapsedGroup: (groupId: string) => void;
+  /** Available for future "dismiss without expanding" UX */
   clearCollapsedGroup: () => void;
   toggleCollapsedGroupVisibility: (groupId: string) => void;
   reset: () => void;
